test(apps/web): Move all transaction tests into a single file
