+++ apps/web/package.json
@@ -13,7 +13,7 @@
         "test:watch": "vitest",
         "test:ui": "vitest --ui",
         "test:ci": "vitest run --coverage",
-        "test:e2e": "playwright test ./e2e/layout/ ./e2e/pages && xvfb-run npx playwright test ./e2e/transactions/depositEther.spec.ts --project chromium && xvfb-run npx playwright test ./e2e/transactions/rawInputForm.spec.ts --project chromium",
+        "test:e2e": "playwright test ./e2e/layout/ ./e2e/pages && xvfb-run npx playwright test ./e2e/transactions/metamask.spec.ts --project chromium",
         "test:e2e:local": "playwright test",
         "test:e2e:ui": "playwright test --ui"
     },
